Updated DeleteInstance Operation
