+++ src/Client/Service.Model/OperationStatus.Context.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OnlineManagementApiClient.Service.Model
+{
+    public class Context
+    {
+        public Dictionary<string, object> Items { get; set; }
+    }
+}
