+++ src/Client/Service.Model/OperationStatus.ItemDescription.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OnlineManagementApiClient.Service.Model
+{
+    public class ItemDescription
+    {
+        public string Subject { get; set; }
+        public string Description { get; set; }
+    }
+}
